+++ src/app/store/slices/storage/fileUtils/prepareFilesToUpload.ts
