+++ src/app/store/slices/storage/folderUtils/checkFolderDuplicated.test.ts
