+++ src/app/store/slices/storage/folderUtils/checkFolderDuplicated.ts
