+++ src/app/store/slices/storage/folderUtils/getUniqueFolderName.test.ts
