+++ src/app/store/slices/storage/folderUtils/getUniqueFolderName.ts
