+++ src/app/store/slices/storage/storage.thunks/renameItemsThunk.ts
@@ -12,10 +12,10 @@ import { SdkFactory } from '../../../../core/factory/sdk';
 import errorService from '../../../../core/services/error.service';
 import { uiActions } from '../../ui';
 import workspacesSelectors from '../../workspaces/workspaces.selectors';
-import { checkDuplicatedFiles } from '../fileNameUtils/checkDuplicatedFiles';
-import { getUniqueFilename } from '../fileNameUtils/getUniqueFilename';
-import { checkFolderDuplicated } from '../folderNameUtils/checkFolderDuplicated';
-import renameFolderIfNeeded from '../folderNameUtils/renameFolderIfNeeded';
+import { checkDuplicatedFiles } from '../fileUtils/checkDuplicatedFiles';
+import { getUniqueFilename } from '../fileUtils/getUniqueFilename';
+import { checkFolderDuplicated } from '../folderUtils/checkFolderDuplicated';
+import renameFolderIfNeeded from '../folderUtils/renameFolderIfNeeded';
 import { StorageState } from '../storage.model';
 import storageSelectors from '../storage.selectors';
 import { IRoot } from './uploadFolderThunk';
